+++ internal/storage/models.go
@@ -16,14 +16,14 @@ type Order struct {
 type Balance struct {
 	ID        int64
 	UserID    int64
-	Amount    int64
-	Withdrawn int64
+	Amount    float32
+	Withdrawn float32
 }
 
 type Transaction struct {
 	ID          int64
 	UserID      int64
 	OrderID     int64
-	Amount      int64
+	Amount      float32
 	ProcessedAt time.Time
 }
